pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["Viktoria Andrushenka <vika.sapriko.9981@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.4"
pandas = "^2.3.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
pillow = "^11.2.1"
django = "^5.2.3"
redis = "^6.2.0"
ipython = "^9.3.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
coverage = "^7.9.1"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"
celery = "^5.5.3"
eventlet = "^0.40.1"
django-celery-beat = "^2.8.1"
drf-spectacular = "^0.28.0"
python-telegram-bot = "20.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.16.1"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'


[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''


[tool.isort]
line_length = 119